backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
python-dotenv==1.0.1
duckdb==1.2.1
anthropic==0.79.0
